Enable determination of layout space by vertex count
Add a construction method for both layout classes that allows for the
determination of the layout space simply by providing the number of
vertices in the graph. the constant k is then set to 100 and the
height/width of the layout space are derived from those to datapoints.
This method will be the default from now on